json struct in extra file and resp. rearrangement of code
