+++ src/neighbourhood_managers/neighbourhood_manager_chain.hh
@@ -52,12 +52,7 @@
 // Each actual implementation of a NeighbourhoodManager is based
 // on the given interface
 #include "neighbourhood_managers/neighbourhood_manager_base.hh"
-#include "neighbourhood_managers/property.hh"
-
-// An external header-library/header-class, which makes it easy to use
-// the JSON as a first class data type. See
-// https://github.com/nlohmann/json for documentation.
-#include "json.hpp"
+#include "neighbourhood_managers/json_io.hh"
 
 // Some data types and operations are based on the Eigen library
 #include <Eigen/Dense>
@@ -66,60 +61,10 @@
 #include <stdexcept>
 #include <vector>
 
-// For convenience
-using json = nlohmann::json;
-
 // All functions and classes are in the namespace <code>rascal</code>,
 // which ensures that they don't clash with other libraries one might
 // use in conjunction.
 namespace rascal {
-
-  //
-  namespace JSONTransfer {
-
-    // To read from a JSON file and deserialize the content, the used
-    // class needs a <code>struct</code> with standard data types.
-    struct AtomicStructure {
-      /**
-	 \param cell is a vector a vector of vectors which holds the cell unit
-	 vectors.
-	 \param type a vector of integers which holds the atomic type
-	 (coordination number).
-	 \param pbc is a 0/1 vector which says, where periodic boundary
-	 conditions are applied.
-	 \param position is a vector of vectors which holds the atomic
-	 positions.
-       */
-      std::vector<std::vector<double>> cell{};
-      std::vector<int> type{};
-      std::vector<int> pbc{};
-      std::vector<std::vector<double>> position{};
-    };
-
-    // This function is used to convert to the JSON format with the
-    // given keywords. It is an overload of the function defined in the header
-    // class json.hpp.
-    // Inline needed, otherwise it is a multiple definition
-    inline void to_json(json & j, AtomicStructure& s) {
-      j = json{
-	{"cell", s.cell},
-	{"numbers", s.type},
-	{"pbc", s.pbc},
-	{"positions", s.position}
-      };
-    }
-
-    // This function is used to read from the JSON file and convert
-    // the data into standard types. It is an overload of the function defined
-    // in json.hpp class header.
-    inline void from_json(const json& j, AtomicStructure& s) {
-      s.cell = j.at("cell").get<std::vector<std::vector<double>>>();
-      s.type = j.at("numbers").get<std::vector<int>>();
-      s.pbc = j.at("pbc").get<std::vector<int>>();
-      s.position = j.at("positions").get<std::vector<std::vector<double>>>();
-    }
-  }
-
   //! forward declaration for traits
   class NeighbourhoodManagerChain;
 
@@ -351,7 +296,7 @@ namespace rascal {
     // <code>atoms_object</code>, because ASE (see above) calls it's
     // read/write function an iterator for 'Atoms objects'.  It is
     // first class C++ data structure, which 'feels' like JSON.
-    JSONTransfer::AtomicStructure atoms_object{};
+    json_io::AtomicStructure atoms_object{};
 
     // Since the data from the <code>atoms_object</code>, especially
     // the positions are not contiguous in memory (they are
